+++ packages/runtime-sdk/tests/web-frameworks-test/flask-tests/src/worker.py
@@ -1,0 +1,5 @@
+from testlib.entrypoint import TestRunnerEntrypoint
+
+
+class Default(TestRunnerEntrypoint):
+    pass